New file structure made
