+++ commands/telegram/tCovid.js
