+++ commands/template.js
@@ -1,0 +1,29 @@
+// DISCORD VERSION
+
+const Discord = require('discord.js');
+const { Telegraf } = require('telegraf')
+
+// Import any extra libraries here
+
+module.exports = {
+    name: '',
+    description: '',
+    async execute(msg, args, Discord, bot) {
+        // Add code here
+    }
+}
+
+/// TELEGRAM VERSION
+
+const Discord = require('discord.js');
+const { Telegraf } = require('telegraf')
+
+// Import any extra libraries here
+
+module.exports = {
+    name: '',
+    description: '',
+    async execute(ctx, Telegraf, telegramBot) {
+        // Add code here
+    }
+}
